Lab 2 Adapters, completed lab and fixed error with add button
