+++ app/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## 1.0.1 - 24-10-2019 ***(Merged)***
+## 1.2.0 - 25-10-2019 ***(Merged)***
+  - TASK - [MOB-DEV-2]  - Lab 2 Adapters, completed and adapted for archaeological hillforts project.
+
+
+## 1.1.0 - 24-10-2019 ***(Merged)***
   - TASK - [MOB-DEV-1]  - Lab 1 Activities, completed and adapted for archaeological hillforts project.
                         - Changelog added.
