+++ app/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme" >
-        <activity android:name="activities.HillfortActivity">
+        <activity android:name=".activities.HillfortActivity">
         </activity>
         <activity android:name=".activities.HillfortListActivity">
             <intent-filter>
