+++ app/src/main/java/app/hillforts/activities/HillfortAdapter.kt
@@ -1,0 +1,31 @@
+package app.hillforts.activities
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import app.hillforts.R
+import app.hillforts.models.HillfortModel
+import kotlinx.android.synthetic.main.card_hillfort.view.*
+
+class HillfortAdapter constructor(private var hillforts: List<HillfortModel>) : RecyclerView.Adapter<HillfortAdapter.MainHolder>() {
+
+  override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MainHolder {
+    return MainHolder(LayoutInflater.from(parent?.context).inflate(R.layout.card_hillfort, parent, false))
+  }
+
+  override fun onBindViewHolder(holder: MainHolder, position: Int) {
+    val hillfort = hillforts[holder.adapterPosition]
+    holder.bind(hillfort)
+  }
+
+  override fun getItemCount(): Int = hillforts.size
+
+  class MainHolder constructor(itemView: View) : RecyclerView.ViewHolder(itemView) {
+
+    fun bind(hillfort: HillfortModel) {
+      itemView.hillfortTitle.text = hillfort.title
+      itemView.description.text = hillfort.description
+    }
+  }
+}
