+++ app/src/main/java/app/hillforts/activities/HillfortListActivity.kt
@@ -4,15 +4,10 @@ import android.os.Bundle
 import android.view.*
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import kotlinx.android.synthetic.main.activity_hillfort_list.*
 import org.jetbrains.anko.startActivityForResult
 import app.hillforts.R
 import app.hillforts.main.MainApp
-import app.hillforts.models.HillfortModel
-import kotlinx.android.synthetic.main.activity_hillfort.view.*
-import kotlinx.android.synthetic.main.card_hillfort.view.description
-import kotlinx.android.synthetic.main.card_hillfort.view.hillfortTitle
 
 class HillfortListActivity : AppCompatActivity() {
 
@@ -40,27 +35,5 @@ class HillfortListActivity : AppCompatActivity() {
             R.id.item_add -> startActivityForResult<HillfortActivity>(0)
         }
         return super.onOptionsItemSelected(item)
-    }
-}
-
-class HillfortAdapter constructor(private var hillforts: List<HillfortModel>) : RecyclerView.Adapter<HillfortAdapter.MainHolder>() {
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MainHolder {
-        return MainHolder(LayoutInflater.from(parent?.context).inflate(R.layout.card_hillfort, parent, false))
-    }
-
-    override fun onBindViewHolder(holder: MainHolder, position: Int) {
-        val hillfort = hillforts[holder.adapterPosition]
-        holder.bind(hillfort)
-    }
-
-    override fun getItemCount(): Int = hillforts.size
-
-    class MainHolder constructor(itemView: View) : RecyclerView.ViewHolder(itemView) {
-
-        fun bind(hillfort: HillfortModel) {
-            itemView.hillfortTitle.text = hillfort.title
-            itemView.description.text = hillfort.description
-        }
     }
 }
